ERR: Was not producing CellData. Reported by Paul Hsieh@usgs.
The filters vtkExtractVOI and vtkExtractGrid appear to
lose cell data. This is illustrated by the attached
sample code. A structured points data set is created
with point and cell data (scalars). After filtering with
vtkExtractVOI, there are point scalars but no cell
scalars. The same thing happens for a structured grid
data set that is extracted by vtkExtractGrid (sample
code not attached).

As always, many thanks for you help and best regards,
Paul