ENH: Enhancements to undo-redo framework. Split vtkSMUndoStack into vtkSMUndoStack and vtkSMUndoStackBuilder. vtkSMUndoStack now only keeps the stack, the generation of undo elements has moved to vtkSMUndoStackBuilder. pqUndoStackBuilder extends vtkSMUndoStackBuilder to automatically add undo elements from property modifications. Added server new undo elements for splitting views etc etc. Undo redo now works for split views, animation, change inputs, application settings dialogs, custom filters etc.
pqPipelineBuilder has been deprecated. It is replaced by pqObjectBuilder. pqObjectBuilder does not have any undo-redo logic. Marking of undo blocks is now the reponsibility of the outermost application layers.

Added SetDefaultValues to several domains. This led to the CLip filter getting confused and using the clip by scalar logic always. For now disabling that functionality in the Clip filter proxy.