BUG: Fixed BUG #4838. "builtin" is not shown in the server connection dialog anymore. This is done by specifying an optional ignore list to the server browser dialog.
Also fixed the bug due to which if the server connection failed while connecting
to a server mentioned in the recent files menu, no default builtin connection was
created.

Also, when openening a recent resource, and multiple configurations for a server
exist, the server browser no longer shows all server resources, instead only the
ones matching the server on which the recent file exists are shown.