ENH: ParaView no longer uses vtkAnimationScene, instead created new vtkPVAnimationScene which is directly a subclass of vtkAnimationCue which uses animation players (vtkAnimationPlayer) subclasses to play the animation separating the scene playing logic from the scene.
Now the time for each tick given by the vtkPVAnimationScene is indeed the
animation time irrespective of play mode (unlike before when the
StartTime/EndTime changed based on play mode).

ParaView now uses cue proxies that define the manipulator as a subproxy making
it easier to create cues. Added a new cue for pipeline time which is added by
default to animate the time keeper's time property. This cue has a mode in which
all keframes are disabled, instead it directly uses the animation clock time to
update the animated property.

ENH: vtkAnimationCue provides information about the animation clock time in each
tick.