BUG: Fixes BUG #5928. This fix entails some explanation. First lets try to understand why the data was always collected on the client unless there view was already compositing. The view needed data size information to decide if compositing should be used. Before getting the size, the view needed to update all representations to ensure that correct data information is used for this check. Also, this data information was obtained from the last UpdateSuppressor i.e. the one that did the caching, so that even when caching was enabled and cache was used, we'd get the correct data information. Hence, since the view updated all representations before calling BeginStillRender or BeginInteractiveRender, all representations used the most recent compositing state to deliver data.
This is now changed. BeginStillRender or BeginInteractiveRender no longer can
assume that all representation will be updated before these methods are
called. Infact, now, the view guarantees only one UpdateAllRepresentations()
call which is after the Begin*Render() and before PerformRender(). All
representations now guarantee that is GetRepresentedDataInformation() or
the methods returning memory sizes are called, the representation will update
the pipeline until before the data movement happens. For this we added a new
PreCollectUpdateSuppressor where-ever required. Thus, when Begin*Render method
does a size query, all representations are "partially" updated and correct
data information is returned. UpdateDataInformation on the representation
strategy updates the pipeline until the data distribution however, when using
cache if the data is cached for the current CacheTime, then it does not update
the pipeline, instead goes to the update suppressor keeping the cache, does a
cache update and then uses the data information from it.