ENH: Added new vtkWeakPointer which implements a weak reference to a vtkObject.
In applications that used textures, it would segfault at times since the render window
reference pointed by the vtkOpenGLTexture object was destroyed before the
texture and the texture had a dangling reference to an obsolete render window.
This issue is overcome by using the new weak reference instead of a generic
pointer.