ENH: Fixing the suggested view type business.
* vtkSMViewProxy has a virtual const char* GetSuggestedViewType(vtkIdType
  * connectionId)
Default implementation simply returns the XMLName for the view proxy itself
(which works for most views such as plots, speadsheet etc).

* vtkSMRenderViewProxy, vtkSMTwoDRenderViewProxy, vtkSMComparativeViewProxy
overrides this method to return appropriate XML name based on the connection
type for the connection ID passed in as an argument.

* pqStandardViewModules::createViewProxy() is the method used by the GUI to
instantiate a view proxy. This method no longer simply creates the proxy for
the xmlname specified. Instead, it create a prototype proxy for the xmlname
specified, then call GetSuggestedViewType() on the prototype to obtain the
real XMLName for the proxy to create and then creates that proxy.

* Same thing is done by the vtkSMPQStateLoader. When a view proxy is requested
to be created, it first creates a prototype proxy using the requested xmlname
and the obtains the suggested XML name.