ENH: Major update to the Python module.
The simple module is going to be the new public interface for ParaView.
The goal behind this module is to provide a simple interface that is
as close to the way the GUI does things as possible. To that end, I
made numerous modifications for both the simple module and the servermanager
module.