Added support for QFileOpenEvent
Handle the QFileOpenEvent which allows a Mac OS X user to double-click a
data file (or use the "open" command). To this end the
pqViewManager::event member function was extended to handle
QEvent::FileOpen which emits the new signal
pqMultiView::triggerOpenFile. The signal is connected to the
pqMainWindowCore::onFileOpen slot in the
pqClientMainWindow::constructorHelper() function.

The implementation opens every file as a single file, never as a
sequence. If a sequence should be opened (e.g. a series of .vpd or .vtk
files), use a .pvd master file. This limitation is due to the fact, that
QFileOpenEvent gets issued for every single file.

Signed-off-by: Michael Wild <themiwi@users.sourceforge.net>