Added the new parallel OpenFOAM reader by Takuya Oshima
Renamed the VTK reader class to vtkNewOpenFOAMReader and keep it in a seperate
directory under VTK/IO because of the Braid submodule. The parallel
reader plugin resides in Plugins/POpenFOAMReader. Some adjustments were
necessary in the CMake code to make the new plugin optional and to
accomodate the non-standard subdirectory of the
vtkNewOpenFOAMReader.{cxx,h} files.

Also the readers.xml file is now configured because the OpenFOAM reader
class name is not fixed any more.

Signed-off-by: Michael Wild <themiwi@users.sourceforge.net>