BUG: Fixed BUG #8658.
In ParaView, we are using a vtkPVSliceFilter to slice the input image data. Now
this filter was requesting only the extent for the requested slice from the
input. So when state file was loaded, the input extent would simply be that of
the slice (not the whole extent for the data that could be produced by the
source). Hence the slider was not movable. Fixed that by making this
vtkPVSliceFilter always request the complete piece from the input.