BUG: No redirection for vtkSocketCommunicatorHash
If the vtkHashSource binary fails to run we do not want any file named
vtkSocketCommunicatorHash.h to be left on disk so that it will be
created when the problem is fixed later.  In order to achieve this we
cannot use shell redirection to put content in the file, as on error it
writes an empty file.  Instead we tell vtkHashSource to open the output
file itself so that no file is created if it does not run.